continua melhorando as telas.

+++ app/src/main/java/com/example/florida/model/Item.kt
@@ -4,8 +4,8 @@ data class Item (
     val id: Long = 0,
     val description: String,
     val qty: Int,
-    val price: Double,
+    val price: Long,
 ) {
-    val total: Double
+    val total: Long
         get() = qty * price
 }
